Improved 3D simulator view
Added 1in checker grid to get size and origin marker

+++ CNC/3Dview.h
@@ -3,8 +3,8 @@
 
 void OnPaint(HWND hWnd);
 
-bool init3DView( double x, double y );
-void initToolShape( double radius );
+bool init3DView( float x, float y );
+void initToolShape( float radius );
 void update3DView( );
 void resetBlockSurface();
 bool start3DViewer();
